feat: support for JSON/YAML parsable table output
Still leveraging the greatness of table options, this enables output for
machine processing.

It is arguably a "table" output, the idea behind putting it inside the
table class is mainly because it relies on same filtering/enhancing
mechanism. I am open to suggestion for refactoring it into its own
styled output.