Provide proper typings for cli.progress()
Hi! 👋 
      
Firstly, thanks for your work on this project! 🙂

At the moment, `cli.progress()` takes an option argument of `any` and returns `any`. This makes it difficult to validate which options are valid and which operations are available on the result of the function.

There may be a reason that `any` is specified here. But if it was an oversight, this PR should resolve the issue.